# Use precompute optimization for Rails.env.local?